fix(web): split plan restore actions and scope locator
